feat(sdk): export mid-level api with direct access to prover calls
