+++ examples/simple-email-proof/vlayer/package.json
@@ -46,14 +46,14 @@
     "tailwindcss": "^3.4.14",
     "autoprefixer": "^10.4.20",
     "@johanneskares/wallet-mock": "^1.4.1",
-    "@playwright/test": "1.48.0",
+    "@playwright/test": "1.52.0",
     "@types/bun": "^1.1.6",
     "@types/eslint__js": "^8.42.3",
     "@types/react": "^18.3.11",
     "@types/react-dom": "^18.3.1",
     "@vitejs/plugin-react": "^4.3.3",
     "globals": "^15.11.0",
-    "playwright": "1.48.0",
+    "playwright": "1.52.0",
     "solhint": "5.0.5",
     "typescript": "^5.5.4",
     "vite": "^5.4.9"
