+++ examples/simple-email-proof/vlayer/playwright.config.ts
@@ -27,7 +27,7 @@ export default defineConfig({
   projects: [
     {
       name: "chromium",
-      use: { ...devices["Desktop Chrome"] },
+      use: { ...devices["Desktop Chrome"], channel: "chromium" },
     },
   ],
   // Run your local dev server before starting the tests.
