+++ examples/simple-time-travel/vlayer/package.json
@@ -43,13 +43,13 @@
   },
   "devDependencies": {
     "@johanneskares/wallet-mock": "^1.4.1",
-    "@playwright/test": "1.48.0",
+    "@playwright/test": "1.52.0",
     "@types/bun": "^1.1.6",
     "@types/react": "^18.3.11",
     "@types/react-dom": "^18.3.1",
     "@vitejs/plugin-react": "^4.3.3",
     "daisyui": "^4.12.14",
-    "playwright": "1.48.0",
+    "playwright": "1.52.0",
     "react-router": "^7.1.5",
     "solhint": "5.0.5",
     "typescript": "^5.5.4",
