+++ examples/simple-time-travel/vlayer/playwright.config.ts
@@ -30,7 +30,7 @@ export default defineConfig({
   projects: [
     {
       name: "chromium",
-      use: { ...devices["Desktop Chrome"] },
+      use: { ...devices["Desktop Chrome"], channel: "chromium" },
     },
   ],
   // Run your local dev server before starting the tests.
