+++ examples/simple-web-proof/vlayer/package.json
@@ -50,12 +50,12 @@
     "@types/node": "^22.5.4",
     "vite-tsconfig-paths": "^5.0.1",
     "@johanneskares/wallet-mock": "^1.4.1",
-    "@playwright/test": "1.48.0",
+    "@playwright/test": "1.52.0",
     "@types/bun": "^1.1.6",
     "@types/chrome": "^0.0.313",
     "@types/debug": "^4.1.12",
     "@vitejs/plugin-react": "^4.3.3",
-    "playwright": "1.48.0",
+    "playwright": "1.52.0",
     "solhint": "5.0.5",
     "vite": "^5.4.1",
     "autoprefixer": "^10.4.20",
