+++ examples/simple-web-proof/vlayer/playwright.config.ts
@@ -29,7 +29,7 @@ export default defineConfig({
   projects: [
     {
       name: "chromium",
-      use: { ...devices["Desktop Chrome"] },
+      use: { ...devices["Desktop Chrome"], channel: "chromium" },
     },
   ],
   // Run your local dev server before starting the tests.
