+++ packages/browser-extension/package.json
@@ -14,7 +14,7 @@
     "test:unit": "vitest --run"
   },
   "devDependencies": {
-    "@playwright/test": "1.48.0",
+    "@playwright/test": "1.52.0",
     "@testing-library/jest-dom": "^6.6.3",
     "@testing-library/react": "^16.0.1",
     "@testing-library/user-event": "^14.5.2",
@@ -43,7 +43,7 @@
     "fp-ts": "^2.16.9",
     "framer-motion": "^11.11.1",
     "jose": "^6.0.10",
-    "playwright": "1.48.0",
+    "playwright": "1.52.0",
     "ramda": "^0.30.1",
     "react": "^18.3.1",
     "react-dom": "^18.3.1",
