+++ packages/sdk/src/api/prover/index.ts
@@ -1,0 +1,4 @@
+export * from "./prover";
+export * from "./v_call";
+export * from "./v_getProofReceipt";
+export * from "./v_versions";
