+++ packages/sdk/src/api/prover/v_getProofReceipt.ts
@@ -1,46 +1,38 @@
-import {
-  type CallContext,
-  type CallParams,
-  type CallHash,
-  callHashSchema,
-} from "types/vlayer";
+import { type CallHash } from "types/vlayer";
 import {
   handleProverResponseError,
   handleAuthErrors,
-} from "./lib/handleErrors";
-import { InvalidProverResponseError } from "./lib/errors";
-import { validateJrpcResponse } from "./lib/jrpc";
+} from "../lib/handleErrors";
+import { InvalidProverResponseError } from "../lib/errors";
+import { proofReceiptSchema, type ProofReceipt } from "../lib/types/vlayer";
+import { validateJrpcResponse } from "../lib/jrpc";
 import debug from "debug";
 
-const log = debug("vlayer:v_call");
+const log = debug("vlayer:v_getProofReceipt");
 
-function v_callBody(call: CallParams, context: CallContext) {
-  log("call", call);
-  log("context", context);
+function v_getProofReceiptBody(hash: CallHash) {
   return {
-    method: "v_call",
-    params: [call, context],
+    method: "v_getProofReceipt",
+    params: { hash },
     id: 1,
     jsonrpc: "2.0",
   };
 }
 
-export async function v_call(
-  call: CallParams,
-  context: CallContext,
+export async function v_getProofReceipt(
+  hash: CallHash,
   url: string = "http://127.0.0.1:3000",
   token?: string,
-): Promise<CallHash> {
+): Promise<ProofReceipt> {
   const headers: Record<string, string> = {
     "Content-Type": "application/json",
   };
   if (token !== undefined) {
     headers["Authorization"] = "Bearer " + token;
   }
-
   const rawResponse = await fetch(url, {
     method: "POST",
-    body: JSON.stringify(v_callBody(call, context)),
+    body: JSON.stringify(v_getProofReceiptBody(hash)),
     headers,
   });
   log("raw response: ", rawResponse);
@@ -58,9 +50,9 @@ export async function v_call(
     throw handleProverResponseError(response.error);
   }
 
-  const result = callHashSchema.safeParse(response.result);
+  const result = proofReceiptSchema.safeParse(response.result);
   if (!result.success) {
-    throw new InvalidProverResponseError("v_call", response.result);
+    throw new InvalidProverResponseError("v_getProofReceipt", response.result);
   }
 
   return result.data;
